requirements.txt
ultralytics
easyocr
pytesseract
streamlit
opencv-python-headless
numpy
Pillow
